#27 Feature: Binance api 연결

+++ src/main/java/com/memesphere/search/dto/response/SearchListPreviewResponse.java
@@ -23,7 +23,7 @@ public class SearchListPreviewResponse {
     @Schema(description = "차트 데이터의 weighted average price", example = "10000")
     BigDecimal weightedAveragePrice;
     @Schema(description = "차트 데이터의 volume", example = "5")
-    Integer volume;
+    BigDecimal volume;
     @Schema(description = "차트 데이터의 price_change", example = "500")
     BigDecimal priceChange;
     @Schema(description = "차트 데이터의 price_change_rate", example = "+2.4%")
